Refactor App.tsx into modular architecture
  - Extract types into separate types.ts file
  - Move i18n translations to dedicated i18n.ts module
  - Separate question data into data/questions.ts
  - Create reusable utility functions in utils.ts
  - Extract custom hooks into hooks.ts module
  - Extract QuestionCard into separate component
  - Improve UX: clickable question text toggles checkbox
  - Fix sidebar scroll to account for sticky header
  - Reduce App.tsx from 700+ lines to ~130 lines
