+++ src/i18n.ts
@@ -1,0 +1,46 @@
+import type { Language, Translation } from './types';
+
+export const I18N: Record<Language, Translation> = {
+  en: {
+    appTitle: 'AI Growth Consultation Checklist',
+    searchPlaceholder: 'Search questions…',
+    clientPlaceholder: 'Client name / session',
+    toList: 'List mode',
+    toFocus: 'Focus mode',
+    json: 'JSON',
+    csv: 'CSV',
+    reset: 'Reset',
+    markDone: 'Mark done',
+    prev: 'Prev',
+    next: 'Next',
+    notAnswered: 'Not answered yet',
+    updated: 'Updated',
+    notesPlaceholder: 'Notes / key points…',
+    complete: 'complete',
+    footer: 'FJ Tech · Consultation Toolkit · Local storage only · Press ←/→ in focus mode',
+    yes: 'Yes',
+    no: 'No',
+    unknown: 'Unknown',
+  },
+  ru: {
+    appTitle: 'Чеклист консультации по росту с ИИ',
+    searchPlaceholder: 'Поиск по вопросам…',
+    clientPlaceholder: 'Имя клиента / сессия',
+    toList: 'Список',
+    toFocus: 'Фокус‑режим',
+    json: 'JSON',
+    csv: 'CSV',
+    reset: 'Сброс',
+    markDone: 'Отметить выполненным',
+    prev: 'Назад',
+    next: 'Далее',
+    notAnswered: 'Пока нет ответа',
+    updated: 'Обновлено',
+    notesPlaceholder: 'Заметки / ключевые моменты…',
+    complete: 'завершено',
+    footer: 'FJ Tech · Инструмент консультаций · Только локальное хранилище · В фокус‑режиме ←/→',
+    yes: 'Да',
+    no: 'Нет',
+    unknown: 'Не знаю',
+  },
+};
